requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-decouple==3.8
dj-database-url==2.3.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
yfinance==0.2.51
pandas==2.3.1
numpy==2.2.6
matplotlib==3.10.7
